@@ -35,12 +35,10 @@ public class MeshNetworkManager: NetworkParametersProvider {
     /// Mesh Network data.
     private var meshData: MeshData
     /// The Network Layer handler.
-    private var networkManager: NetworkManager?
+    internal private(set) var networkManager: NetworkManager?
     /// Storage to keep the app data.
     private let storage: Storage
     
-    /// A queue to handle incoming and outgoing messages.
-    internal let queue: DispatchQueue
     /// A queue to call delegate methods on.
     internal let delegateQueue: DispatchQueue
     
@@ -99,30 +97,25 @@ public class MeshNetworkManager: NetworkParametersProvider {
     ///
     /// - parameters:
     ///   - storage: The storage to use to save the network configuration.
-    ///   - queue:   The `DispatchQueue` to process requests on. By default, the a global
-    ///              background concurrent queue will be used.
-    ///              Note, that if multiple messages are sent shortly one after another,
-    ///              processing them in a concurrent queue may cause some of them to be
-    ///              discarded despite the fact that they were received in the ascending
-    ///              order of SeqAuth, as one with a greater SeqAuth value may be processed
-    ///              before the previous one, causing the replay protection validation fail
-    ///              for the latter. This library stores 2 last SeqAuth values, so if a
-    ///              message with a unique SeqAuth is processed after its successor, it
-    ///              will be processed correctly.
     ///   - delegateQueue: The `DispatchQueue` to call delegate methods on.
     ///                    By default the global main queue will be used.
     /// - seeAlso: ``LocalStorage``
     /// - seeAlso: ``LowerTransportLayer.checkAgainstReplayAttack(_:NetworkPdu)``
     public init(using storage: Storage = LocalStorage(),
-                queue: DispatchQueue = DispatchQueue.global(qos: .background),
                 delegateQueue: DispatchQueue = DispatchQueue.main) {
         self.storage = storage
         self.meshData = MeshData()
-        self.queue = queue
         self.delegateQueue = delegateQueue
         self.proxyFilter = ProxyFilter(delegateQueue)
         // Only now self can be used.
         self.proxyFilter.use(with: self)
+    }
+    
+    @available(*, deprecated, renamed: "init(using:delegateQueue:)")
+    public convenience init(using storage: Storage = LocalStorage(),
+                queue: DispatchQueue = DispatchQueue.global(qos: .background),
+                delegateQueue: DispatchQueue = DispatchQueue.main) {
+        self.init(using: storage, delegateQueue: delegateQueue)
     }
     
     /// Initializes the Mesh Network Manager. It will use the ``LocalStorage`` with the given
@@ -135,25 +128,21 @@ public class MeshNetworkManager: NetworkParametersProvider {
     ///
     /// - parameters:
     ///   - fileName: File name to keep the configuration.
-    ///   - queue:    The `DispatchQueue` to process requests on. By default, the a global
-    ///               background concurrent queue will be used.
-    ///               Note, that if multiple messages are sent shortly one after another,
-    ///               processing them in a concurrent queue may cause some of them to be
-    ///               discarded despite the fact that they were received in the ascending
-    ///               order of SeqAuth, as one with a greater SeqAuth value may be processed
-    ///               before the previous one, causing the replay protection validation fail
-    ///               for the latter. This library stores 2 last SeqAuth values, so if a
-    ///               message with a unique SeqAuth is processed after its successor, it
-    ///               will be processed correctly.
     ///   - delegateQueue: The `DispatchQueue` to call delegate methods on.
     ///                    By default the global main queue will be used.
     ///
     /// - seeAlso: ``LocalStorage``
     public convenience init(using fileName: String,
+                            delegateQueue: DispatchQueue = DispatchQueue.main) {
+        self.init(using: LocalStorage(fileName: fileName),
+                  delegateQueue: delegateQueue)
+    }
+    
+    @available(*, deprecated, renamed: "init(using:delegateQueue:)")
+    public convenience init(using fileName: String,
                             queue: DispatchQueue = DispatchQueue.global(qos: .background),
                             delegateQueue: DispatchQueue = DispatchQueue.main) {
         self.init(using: LocalStorage(fileName: fileName),
-                  queue: queue,
                   delegateQueue: delegateQueue)
     }
     
@@ -276,7 +265,7 @@ public extension MeshNetworkManager {
         guard let networkManager = networkManager else {
             return
         }
-        queue.async {
+        Task.detached {
             networkManager.handle(incomingPdu: data, ofType: type)
         }
     }
@@ -310,7 +299,7 @@ public extension MeshNetworkManager {
               let _ = meshNetwork?.applicationKeys[publish.index] else {
             return nil
         }
-        queue.async {
+        Task {
             networkManager.publish(message, from: model)
         }
         return MessageHandle(for: message, sentFrom: localElement.unicastAddress,
@@ -319,6 +308,8 @@ public extension MeshNetworkManager {
     
     /// Encrypts the message with the Application Key and the Network Key
     /// bound to it, and sends to the given destination address.
+    ///
+    /// The method completes when the message has been sent or an error occurred.
     ///
     /// An appropriate callback of the ``MeshNetworkDelegate`` will be called when
     /// the message has been sent successfully or a problem occured.
@@ -332,19 +323,15 @@ public extension MeshNetworkManager {
     ///   - initialTtl:     The initial TTL (Time To Live) value of the message.
     ///                     If `nil`, the default Node TTL will be used.
     ///   - applicationKey: The Application Key to sign the message.
-    ///   - completion:     The completion handler called when the message
-    ///                     has been sent.
     /// - throws: This method throws when the mesh network has not been created,
     ///           the local Node does not have configuration capabilities
     ///           (no Unicast Address assigned), or the given local Element
-    ///           does not belong to the local Node.
-    /// - returns: Message handle that can be used to cancel sending.
-    @discardableResult
+    ///           does not belong to the local Node, or the manager failed to
+    ///           send the message.
     func send(_ message: MeshMessage,
               from localElement: Element? = nil, to destination: MeshAddress,
               withTtl initialTtl: UInt8? = nil,
-              using applicationKey: ApplicationKey,
-              completion: ((Result<Void, Error>) -> ())? = nil) throws -> MessageHandle {
+              using applicationKey: ApplicationKey) async throws {
         guard let networkManager = networkManager,
               let meshNetwork = meshNetwork else {
             print("Error: Mesh Network not created")
@@ -363,17 +350,14 @@ public extension MeshNetworkManager {
             print("Error: TTL value \(initialTtl!) is invalid")
             throw AccessError.invalidTtl
         }
-        queue.async {
-            networkManager.send(message, from: source, to: destination,
-                                withTtl: initialTtl, using: applicationKey,
-                                completion: completion)
-        }
-        return MessageHandle(for: message, sentFrom: source.unicastAddress,
-                             to: destination.address, using: networkManager)
+        try await networkManager.send(message, from: source, to: destination,
+                                      withTtl: initialTtl, using: applicationKey)
     }
     
     /// Encrypts the message with the Application Key and a Network Key
     /// bound to it, and sends to the given ``Group``.
+    ///
+    /// The method completes when the message has been sent or an error occurred.
     ///
     /// An appropriate callback of the ``MeshNetworkDelegate`` will be called when
     /// the message has been sent successfully or a problem occured.
@@ -387,27 +371,24 @@ public extension MeshNetworkManager {
     ///   - initialTtl:     The initial TTL (Time To Live) value of the message.
     ///                     If `nil`, the default Node TTL will be used.
     ///   - applicationKey: The Application Key to sign the message.
-    ///   - completion:     The completion handler called when the message
-    ///                     has been sent.
     /// - throws: This method throws when the mesh network has not been created,
     ///           the local Node does not have configuration capabilities
     ///           (no Unicast Address assigned), or the given local Element
-    ///           does not belong to the local Node.
-    /// - returns: Message handle that can be used to cancel sending.
-    @discardableResult
+    ///           does not belong to the local Node, or the manager failed to
+    ///           send the message.
     func send(_ message: MeshMessage,
               from localElement: Element? = nil, to group: Group,
               withTtl initialTtl: UInt8? = nil,
-              using applicationKey: ApplicationKey,
-              completion: ((Result<Void, Error>) -> ())? = nil) throws -> MessageHandle {
-        return try send(message, from: localElement, to: group.address,
-                        withTtl: initialTtl, using: applicationKey,
-                        completion: completion)
+              using applicationKey: ApplicationKey) async throws {
+        try await send(message, from: localElement, to: group.address,
+                       withTtl: initialTtl, using: applicationKey)
     }
     
     /// Encrypts the message with the first Application Key bound to the given
     /// ``Model`` and the Network Key bound to it, and sends it to the Node
     /// to which the Model belongs to.
+    ///
+    /// The method completes when the message has been sent or an error occurred.
     ///
     /// An appropriate callback of the ``MeshNetworkDelegate`` will be called when
     /// the message has been sent successfully or a problem occured.
@@ -420,20 +401,16 @@ public extension MeshNetworkManager {
     ///   - model:          The destination Model.
     ///   - initialTtl:     The initial TTL (Time To Live) value of the message.
     ///                     If `nil`, the default Node TTL will be used.
-    ///   - completion:     The completion handler called when the message
-    ///                     has been sent.
     /// - throws: This method throws when the mesh network has not been created,
     ///           the target Model does not belong to any Element, or has
     ///           no Application Key bound to it, or when
     ///           the local Node does not have configuration capabilities
     ///           (no Unicast Address assigned), or the given local Element
-    ///           does not belong to the local Node.
-    /// - returns: Message handle that can be used to cancel sending.
-    @discardableResult
+    ///           does not belong to the local Node, or the manager failed to
+    ///           send the message.
     func send(_ message: UnacknowledgedMeshMessage,
               from localElement: Element? = nil, to model: Model,
-              withTtl initialTtl: UInt8? = nil,
-              completion: ((Result<Void, Error>) -> ())? = nil) throws -> MessageHandle {
+              withTtl initialTtl: UInt8? = nil) async throws {
         guard let element = model.parentElement else {
             print("Error: Element does not belong to a Node")
             throw AccessError.invalidDestination
@@ -444,13 +421,15 @@ public extension MeshNetworkManager {
             print("Error: Model is not bound to any Application Key")
             throw AccessError.modelNotBoundToAppKey
         }
-        return try send(message, from: localElement, to: MeshAddress(element.unicastAddress),
-                        withTtl: initialTtl, using: applicationKey, completion: completion)
+        try await send(message, from: localElement, to: MeshAddress(element.unicastAddress),
+                       withTtl: initialTtl, using: applicationKey)
     }
     
     /// Encrypts the message with the common Application Key bound to both given
     /// ``Model``s and the Network Key bound to it, and sends it to the Node
     /// to which the target Model belongs to.
+    ///
+    /// The method completes when the message has been sent or an error occurred.
     ///
     /// An appropriate callback of the ``MeshNetworkDelegate`` will be called when
     /// the message has been sent successfully or a problem occured.
@@ -463,31 +442,27 @@ public extension MeshNetworkManager {
     ///   - model:        The destination Model.
     ///   - initialTtl:   The initial TTL (Time To Live) value of the message.
     ///                   If `nil`, the default Node TTL will be used.
-    ///   - completion:   The completion handler called when the message
-    ///                   has been sent.
     /// - throws: This method throws when the mesh network has not been created,
     ///           the local or target Model do not belong to any Element, or have
     ///           no common Application Key bound to them, or when
     ///           the local Node does not have configuration capabilities
     ///           (no Unicast Address assigned), or the given local Element
-    ///           does not belong to the local Node.
-    /// - returns: Message handle that can be used to cancel sending.
-    @discardableResult
+    ///           does not belong to the local Node, or the manager failed to
+    ///           send the message.
     func send(_ message: UnacknowledgedMeshMessage,
               from localModel: Model, to model: Model,
-              withTtl initialTtl: UInt8? = nil,
-              completion: ((Result<Void, Error>) -> ())? = nil) throws -> MessageHandle {
+              withTtl initialTtl: UInt8? = nil) async throws {
         guard let localElement = localModel.parentElement else {
             print("Error: Source Model does not belong to an Element")
             throw AccessError.invalidSource
         }
-        return try send(message, from: localElement, to: model,
-                        withTtl: initialTtl, completion: completion)
+        try await send(message, from: localElement, to: model,
+                       withTtl: initialTtl)
     }
     
     /// Encrypts the message with the first Application Key bound to the given
-    /// Model and a Network Key bound to it, and sends it to the Node
-    /// to which the Model belongs to.
+    /// ``Model`` and a Network Key bound to it, and sends it to the Node
+    /// to which the Model belongs to and returns the response.
     ///
     /// An appropriate callback of the ``MeshNetworkDelegate`` will be called when
     /// the message has been sent successfully or a problem occured.
@@ -500,21 +475,17 @@ public extension MeshNetworkManager {
     ///   - model:          The destination Model.
     ///   - initialTtl:     The initial TTL (Time To Live) value of the message.
     ///                     If `nil`, the default Node TTL will be used.
-    ///   - applicationKey: The Application Key to sign the message.
-    ///   - completion:     The completion handler called when the response
-    ///                     has been received.
     /// - throws: This method throws when the mesh network has not been created,
     ///           the target Model does not belong to any Element, or has
     ///           no Application Key bound to it, or when
     ///           the local Node does not have configuration capabilities
     ///           (no Unicast Address assigned), or the given local Element
     ///           does not belong to the local Node.
-    /// - returns: Message handle that can be used to cancel sending.
-    @discardableResult
+    /// - returns: The response with the expected ``AcknowledgedMeshMessage/responseOpCode``
+    ///            received from the target Node.
     func send(_ message: AcknowledgedMeshMessage,
               from localElement: Element? = nil, to model: Model,
-              withTtl initialTtl: UInt8? = nil,
-              completion: ((Result<MeshResponse, Error>) -> ())? = nil) throws -> MessageHandle {
+              withTtl initialTtl: UInt8? = nil) async throws -> MeshResponse {
         guard let networkManager = networkManager,
               let meshNetwork = meshNetwork else {
             print("Error: Mesh Network not created")
@@ -542,17 +513,13 @@ public extension MeshNetworkManager {
             print("Error: TTL value \(initialTtl!) is invalid")
             throw AccessError.invalidTtl
         }
-        queue.async {
-            networkManager.send(message, from: source, to: element.unicastAddress,
-                                withTtl: initialTtl, using: applicationKey,
-                                completion: completion)
-        }
-        return MessageHandle(for: message, sentFrom: source.unicastAddress,
-                             to: element.unicastAddress, using: networkManager)
+        return try await networkManager
+            .send(message, from: source, to: element.unicastAddress,
+                  withTtl: initialTtl, using: applicationKey)
     }
     
     /// Encrypts the message with the common Application Key bound to both given
-    /// Models and a Network Key bound to it, and sends it to the Node
+    /// ``Model``s and a Network Key bound to it, and sends it to the Node
     /// to which the target Model belongs to.
     ///
     /// An appropriate callback of the ``MeshNetworkDelegate`` will be called when
@@ -566,29 +533,27 @@ public extension MeshNetworkManager {
     ///   - model:        The destination Model.
     ///   - initialTtl:   The initial TTL (Time To Live) value of the message.
     ///                   If `nil`, the default Node TTL will be used.
-    ///   - completion:   The completion handler which is called when the response
-    ///                   has been received.
     /// - throws: This method throws when the mesh network has not been created,
     ///           the local or target Model do not belong to any Element, or have
     ///           no common Application Key bound to them, or when
     ///           the local Node does not have configuration capabilities
     ///           (no Unicast Address assigned), or the given local Element
     ///           does not belong to the local Node.
-    /// - returns: Message handle that can be used to cancel sending.
-    @discardableResult
+    /// - returns: The response with the expected ``AcknowledgedMeshMessage/responseOpCode``
+    ///            received from the target Node.
     func send(_ message: AcknowledgedMeshMessage,
               from localModel: Model, to model: Model,
-              withTtl initialTtl: UInt8? = nil,
-              completion: ((Result<MeshResponse, Error>) -> ())? = nil) throws -> MessageHandle {
+              withTtl initialTtl: UInt8? = nil) async throws -> MeshResponse {
         guard let localElement = localModel.parentElement else {
             print("Error: Source Model does not belong to an Element")
             throw AccessError.invalidSource
         }
-        return try send(message, from: localElement, to: model,
-                        withTtl: initialTtl, completion: completion)
+        return try await send(message, from: localElement, to: model,
+                              withTtl: initialTtl)
     }
     
-    /// Sends Configuration Message to the Node with given destination Address.
+    /// Sends a Configuration Message to the Node with given destination address
+    /// and returns the received response.
     ///
     /// The `destination` must be a Unicast Address, otherwise the method
     /// throws an ``AccessError/invalidDestination`` error.
@@ -601,8 +566,6 @@ public extension MeshNetworkManager {
     ///   - destination: The destination Unicast Address.
     ///   - initialTtl:  The initial TTL (Time To Live) value of the message.
     ///                  If `nil`, the default Node TTL will be used.
-    ///   - completion:   The completion handler called when the message
-    ///                   has been sent.
     /// - throws: This method throws when the mesh network has not been created,
     ///           the local Node does not have configuration capabilities
     ///           (no Unicast Address assigned), or the destination address
@@ -610,10 +573,8 @@ public extension MeshNetworkManager {
     ///           Error ``AccessError/cannotDelete`` is sent when trying to
     ///           delete the last Network Key on the device.
     /// - returns: Message handle that can be used to cancel sending.
-    @discardableResult
     func send(_ message: UnacknowledgedConfigMessage, to destination: Address,
-              withTtl initialTtl: UInt8? = nil,
-              completion: ((Result<Void, Error>) -> ())? = nil) throws -> MessageHandle {
+              withTtl initialTtl: UInt8? = nil) async throws {
         guard let networkManager = networkManager,
               let meshNetwork = meshNetwork else {
             print("Error: Mesh Network not created")
@@ -644,12 +605,8 @@ public extension MeshNetworkManager {
             print("Error: TTL value \(initialTtl!) is invalid")
             throw AccessError.invalidTtl
         }
-        queue.async {
-            networkManager.send(message, from: element, to: destination,
-                                withTtl: initialTtl, completion: completion)
-        }
-        return MessageHandle(for: message, sentFrom: element.unicastAddress,
-                             to: destination, using: networkManager)
+        try await networkManager.send(message, from: element, to: destination,
+                                      withTtl: initialTtl)
     }
     
     /// Sends a Configuration Message to the primary Element on the given ``Node``.
@@ -662,8 +619,6 @@ public extension MeshNetworkManager {
     ///   - node:       The destination Node.
     ///   - initialTtl: The initial TTL (Time To Live) value of the message.
     ///                 If `nil`, the default Node TTL will be used.
-    ///   - completion: The completion handler which is called when the response
-    ///                 has been received.
     /// - throws: This method throws when the mesh network has not been created,
     ///           the local Node does not have configuration capabilities
     ///           (no Unicast Address assigned), or the destination address
@@ -671,12 +626,10 @@ public extension MeshNetworkManager {
     ///           Error ``AccessError/cannotDelete`` is sent when trying to
     ///           delete the last Network Key on the device.
     /// - returns: Message handle that can be used to cancel sending.
-    @discardableResult
     func send(_ message: UnacknowledgedConfigMessage, to node: Node,
-              withTtl initialTtl: UInt8? = nil,
-              completion: ((Result<Void, Error>) -> ())? = nil) throws -> MessageHandle {
-        return try send(message, to: node.primaryUnicastAddress,
-                        withTtl: initialTtl, completion: completion)
+              withTtl initialTtl: UInt8? = nil) async throws {
+        return try await send(message, to: node.primaryUnicastAddress,
+                              withTtl: initialTtl)
     }
     
     /// Sends Configuration Message to the Node with given destination Address.
@@ -692,19 +645,16 @@ public extension MeshNetworkManager {
     ///   - destination: The destination Unicast Address.
     ///   - initialTtl:  The initial TTL (Time To Live) value of the message.
     ///                  If `nil`, the default Node TTL will be used.
-    ///   - completion:  The completion handler which is called when the response
-    ///                  has been received.
     /// - throws: This method throws when the mesh network has not been created,
     ///           the local Node does not have configuration capabilities
     ///           (no Unicast Address assigned), or the destination address
     ///           is not a Unicast Address or it belongs to an unknown Node.
     ///           Error ``AccessError/cannotDelete`` is sent when trying to
     ///           delete the last Network Key on the device.
-    /// - returns: Message handle that can be used to cancel sending.
-    @discardableResult
-    func send(_ message: AcknowledgedConfigMessage, to destination: Address,
-              withTtl initialTtl: UInt8? = nil,
-              completion: ((Result<ConfigResponse, Error>) -> ())? = nil) throws -> MessageHandle {
+    /// - returns: The response associated with the message.
+    func send(_ message: AcknowledgedConfigMessage,
+              to destination: Address,
+              withTtl initialTtl: UInt8? = nil) async throws -> ConfigResponse {
         guard let networkManager = networkManager,
               let meshNetwork = meshNetwork else {
             print("Error: Mesh Network not created")
@@ -741,15 +691,12 @@ public extension MeshNetworkManager {
             print("Error: TTL value \(initialTtl!) is invalid")
             throw AccessError.invalidTtl
         }
-        queue.async {
-            networkManager.send(message, from: element, to: destination,
-                                withTtl: initialTtl, completion: completion)
-        }
-        return MessageHandle(for: message, sentFrom: element.unicastAddress,
-                             to: destination, using: networkManager)
+        return try await networkManager
+            .send(message, from: element, to: destination, withTtl: initialTtl)
     }
     
-    /// Sends a Configuration Message to the primary Element on the given ``Node``.
+    /// Sends a Configuration Message to the primary Element on the given ``Node``
+    /// and returns the received response.
     ///
     /// An appropriate callback of the ``MeshNetworkDelegate`` will be called when
     /// the message has been sent successfully or a problem occured.
@@ -759,38 +706,35 @@ public extension MeshNetworkManager {
     ///   - node:       The destination Node.
     ///   - initialTtl: The initial TTL (Time To Live) value of the message.
     ///                 If `nil`, the default Node TTL will be used.
-    ///   - completion: The completion handler which is called when the response
-    ///                 has been received.
     /// - throws: This method throws when the mesh network has not been created,
     ///           the local Node does not have configuration capabilities
     ///           (no Unicast Address assigned), or the destination address
     ///           is not a Unicast Address or it belongs to an unknown Node.
     ///           Error ``AccessError/cannotDelete`` is sent when trying to
     ///           delete the last Network Key on the device.
-    /// - returns: Message handle that can be used to cancel sending.
-    @discardableResult
-    func send(_ message: AcknowledgedConfigMessage, to node: Node,
-              withTtl initialTtl: UInt8? = nil,
-              completion: ((Result<ConfigResponse, Error>) -> ())? = nil) throws -> MessageHandle {
-        return try send(message, to: node.primaryUnicastAddress,
-                        withTtl: initialTtl, completion: completion)
+    /// - returns: The response associated with the message.
+    func send(_ message: AcknowledgedConfigMessage,
+              to node: Node,
+              withTtl initialTtl: UInt8? = nil) async throws -> ConfigResponse {
+        return try await send(message, to: node.primaryUnicastAddress,
+                              withTtl: initialTtl)
     }
     
-    /// Sends the Configuration Message to the primary Element of the local Node.
+    /// Sends the Configuration Message to the primary Element of the local ``Node``
+    /// and returns the received response.
     ///
-    /// An appropriate callback of the ``MeshNetworkDelegate`` will be called when
+    /// An appropriate callback of the ``MeshNetworkDelegate`` will also be called when
     /// the message has been sent successfully or a problem occured.
     ///
+    /// - parameters:
     /// - parameter message: The acknowledged configuration message to be sent.
     /// - throws: This method throws when the mesh network has not been created,
-    ///           or the local Node does not have configuration capabilities
-    ///           (no Unicast Address assigned).
+    ///           the local Node does not have configuration capabilities
+    ///           (no Unicast Address assigned) or the local Node returned an error.
     ///           Error ``AccessError/cannotDelete`` is sent when trying to
-    ///           delete the last Network Key on the device.
-    /// - returns: Message handle that can be used to cancel sending.
-    @discardableResult
-    func sendToLocalNode(_ message: AcknowledgedConfigMessage,
-                         completion: ((Result<ConfigResponse, Error>) -> ())? = nil) throws -> MessageHandle {
+    ///           delete the last Network Key on the Node.
+    /// - returns: The response associated with the message.
+    func sendToLocalNode(_ message: AcknowledgedConfigMessage) async throws -> ConfigResponse {
         guard let meshNetwork = meshNetwork else {
             print("Error: Mesh Network not created")
             throw MeshNetworkError.noNetwork
@@ -800,7 +744,7 @@ public extension MeshNetworkManager {
             print("Error: Local Provisioner has no Unicast Address assigned")
             throw AccessError.invalidSource
         }
-        return try send(message, to: destination, withTtl: 1, completion: completion)
+        return try await send(message, to: destination, withTtl: 1)
     }
     
     /// Sends the Proxy Configuration Message to the connected Proxy Node.
@@ -819,8 +763,8 @@ public extension MeshNetworkManager {
             print("Error: Mesh Network not created")
             throw MeshNetworkError.noNetwork
         }
-        queue.async {
-            networkManager.send(message)
+        Task {
+            await networkManager.send(message)
         }
     }
     
@@ -832,7 +776,7 @@ public extension MeshNetworkManager {
             print("Error: Mesh Network not created")
             throw MeshNetworkError.noNetwork
         }
-        queue.async {
+        Task {
             networkManager.cancel(messageId)
         }
     }
