@@ -148,13 +148,17 @@ class AppKeysViewController: UITableViewController, Editable {
         return applicationKey.isUsed(in: network) ? .none : .delete
     }
     
-    override func tableView(_ tableView: UITableView, editActionsForRowAt indexPath: IndexPath) -> [UITableViewRowAction]? {
+    override func tableView(_ tableView: UITableView, trailingSwipeActionsConfigurationForRowAt indexPath: IndexPath) -> UISwipeActionsConfiguration? {
         let network = MeshNetworkManager.instance.meshNetwork!
         let applicationKey = network.applicationKeys[indexPath.keyIndex]
         
         // It should not be possible to delete a key that is in use.
         if applicationKey.isUsed(in: network) {
-            return [UITableViewRowAction(style: .normal, title: "Key in use", handler: {_,_ in })]
+            return UISwipeActionsConfiguration(actions: [
+                UIContextualAction(style: .normal, title: "Key in use", handler: { _, _, completionHandler in
+                    completionHandler(false)
+                })
+            ])
         }
         return nil
     }
