@@ -118,14 +118,18 @@ class NetworkKeysViewController: UITableViewController, Editable {
         return networkKey.isPrimary || networkKey.isUsed(in: network) ? .none : .delete
     }
     
-    override func tableView(_ tableView: UITableView, editActionsForRowAt indexPath: IndexPath) -> [UITableViewRowAction]? {
+    override func tableView(_ tableView: UITableView, trailingSwipeActionsConfigurationForRowAt indexPath: IndexPath) -> UISwipeActionsConfiguration? {
         let network = MeshNetworkManager.instance.meshNetwork!
         let networkKey = network.networkKeys[indexPath.keyIndex]
         
         // It should not be possible to delete a key that is in use.
         if networkKey.isPrimary || networkKey.isUsed(in: network) {
             let title = networkKey.isPrimary ? "Primary Key" : "Key in use"
-            return [UITableViewRowAction(style: .normal, title: title, handler: {_,_ in })]
+            return UISwipeActionsConfiguration(actions: [
+                UIContextualAction(style: .normal, title: title, handler: { _, _, completionHandler in
+                    completionHandler(false)
+                })
+            ])
         }
         return nil
     }
